requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
aiofiles==0.7.0
    # via -r requirements.in
anyio==3.3.2
    # via
    #   httpcore
    #   starlette
arrow==1.2.0
    # via -r requirements.in
asgiref==3.4.1
    # via uvicorn
attrs==21.2.0
    # via -r requirements.in
certifi==2021.10.8
    # via httpx
charset-normalizer==2.0.6
    # via httpx
click==8.0.2
    # via
    #   typer
    #   uvicorn
h11==0.12.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.13.7
    # via httpx
httpx==1.0.0b0
    # via -r requirements.in
idna==3.2
    # via
    #   anyio
    #   rfc3986
python-dateutil==2.8.2
    # via arrow
rfc3986[idna2008]==1.5.0
    # via httpx
six==1.16.0
    # via python-dateutil
sniffio==1.2.0
    # via
    #   anyio
    #   httpcore
    #   httpx
starlette==0.16.0
    # via -r requirements.in
tqdm==4.62.3
    # via -r requirements.in
typer==0.4.0
    # via -r requirements.in
uvicorn==0.15.0
    # via -r requirements.in
